#3 v1.0
Issues Faced.

- As a beginner, I do not know the meaning of Japanese words.
The system didn’t ask at what level I related to Japanese.
Also, we need to come up with levels and what it will be based on as a team.
Do we need to ask the user at what level he/she is?
How the user will be able to decide at what level he/she is, as we need to provide reference based on how we have set the levels?