#3 v.1.3 Pointers
The level was reduced to beginner and elementary, which I am not sure is okay or not. As we are pitching this product as beginner-friendly. So it might work. But a little advanced user may see the initial levels quite easily.
What I liked about the system was teaching sentence construction by letting them construct the English to Japanese directly and then, later on, telling them that they were close and introducing sentence construction steps one by one through quick tips. This was a really good experience from a beginner's standpoint.
At this point, the user is not sure at what stage he/she is and how much he/she has learned. We need to use the gamification method to make them realize that they are how far in completing each stage.

The stages should be linear in difficulty while adjusting themselves if the user is making any mistakes. 
Each stage should either be teaching a concept or a scenario(conversations in restaurants, buses, train tickets, etc.) that they will face while in Japan. We need to have a mix so that they can enjoy each session. 
We need to have a showcase of how many stages will be there. After completing a set of stages, what they have achieved and what they can do while in Japan. Completing all the stages will assist them in completing the JLPT N5 level and traveling more confidently in Japan.
Also initially we need to let them know what the romaji format is as they are beginner-level.